Refactor `labelFunction` and remove `crossAxis`
